fix: move the test for scale_wind_speed in a seperate file
